backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
cryptography==41.0.4
pandas==2.2.3
numpy==2.2.4